# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "nmrXiv",
        "version": "1.0"
    },
    "paths": {
...
        "/api/auth/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Register new user account",
                "description": "Creates a new user account in the NMRXIV platform. Supports both regular user registration and ELN (Electronic Lab Notebook) user registration with different verification flows. Regular users receive email verification, while ELN users are auto-verified with a 3-day welcome period.",
                "operationId": "registerUser",
                "requestBody": {
                    "description": "User registration data including personal information and credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "first_name",
                                    "last_name",
                                    "email",
                                    "password",
                                    "username"
                                ],
                                "properties": {
                                    "first_name": {
                                        "description": "User's first name",
                                        "type": "string",
                                        "example": "Dr. Sarah",
                                        "maxLength": 255,
                                        "minLength": 1
                                    },
                                    "last_name": {
                                        "description": "User's last name",
                                        "type": "string",
                                        "example": "Johnson",
                                        "maxLength": 255,
                                        "minLength": 1
                                    },
                                    "email": {
                                        "description": "User's email address (must be unique)",
                                        "type": "string",
                                        "format": "email",
                                        "example": "sarah.johnson@university.edu",
                                        "maxLength": 255
                                    },
                                    "username": {
                                        "description": "Unique username for the account",
                                        "type": "string",
                                        "example": "sarah_johnson_chem",
                                        "maxLength": 255,
                                        "minLength": 1
                                    },
                                    "password": {
                                        "description": "Password for the account (minimum 8 characters)",
                                        "type": "string",
                                        "format": "password",
                                        "example": "SecurePassword123!",
                                        "minLength": 8
                                    },
                                    "orcid_id": {
                                        "description": "Optional ORCID identifier for academic users",
                                        "type": "string",
                                        "pattern": "^\\d{4}-\\d{4}-\\d{4}-\\d{3}[\\dX]$",
                                        "example": "0000-0002-1825-0097",
                                        "nullable": true
                                    },
                                    "affiliation": {
                                        "description": "Optional institutional affiliation",
                                        "type": "string",
                                        "example": "University of Chemistry, Department of Organic Chemistry",
                                        "nullable": true,
                                        "maxLength": 500
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User registration successful - Account created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "Indicates successful registration",
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "description": "Success message with next steps",
                                            "type": "string",
                                            "example": "User creation successful. Kindly confirm your email address by clicking the link sent to your inbox."
                                        },
                                        "access_token": {
                                            "description": "Bearer token for immediate API access",
                                            "type": "string",
                                            "example": "2|abc123def456ghi789jkl012mno345pqr678stu901vwx234yz"
                                        },
                                        "token_type": {
                                            "description": "Type of the token issued",
                                            "type": "string",
                                            "example": "Bearer"
                                        },
                                        "user_id": {
                                            "description": "ID of the newly created user",
                                            "type": "integer",
                                            "example": 123
                                        },
                                        "verification_required": {
                                            "description": "Whether email verification is required",
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Validation failed - Invalid input data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "Request status",
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "description": "Error type description",
                                            "type": "string",
                                            "example": "validation error"
                                        },
                                        "errors": {
                                            "description": "Field-specific validation errors",
                                            "properties": {
                                                "email": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    },
                                                    "example": [
                                                        "The email has already been taken.",
                                                        "The email must be a valid email address."
                                                    ]
                                                },
                                                "username": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    },
                                                    "example": [
                                                        "The username has already been taken."
                                                    ]
                                                },
                                                "password": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    },
                                                    "example": [
                                                        "The password must be at least 8 characters."
                                                    ]
                                                },
                                                "first_name": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    },
                                                    "example": [
                                                        "The first name field is required."
                                                    ]
                                                },
                                                "last_name": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    },
                                                    "example": [
                                                        "The last name field is required."
                                                    ]
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable entity - Business logic validation failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "description": "Additional validation errors",
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "429": {
                        "description": "Too many registration attempts - Rate limit exceeded",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Too many registration attempts. Please try again later."
                                        },
                                        "retry_after": {
                                            "description": "Seconds until next attempt is allowed",
                                            "type": "integer",
                                            "example": 60
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error during registration",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "An error occurred during user registration."
                                        },
                                        "error_code": {
                                            "type": "string",
                                            "example": "REGISTRATION_FAILED"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "token": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            },
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}